+++ dlls/comdlg32/printdlg.c
@@ -79,7 +79,8 @@ static UINT_PTR CALLBACK print_hook_proc(HWND hdlg, UINT msg, WPARAM wp, LPARAM 
     call = HeapAlloc(GetProcessHeap(), 0, sizeof(*call));
 #endif
 
-    WINE_TRACE("Calling guest proc 0x%lx(%p, %x, %lu, %lu).\n", *guest_proc, hdlg, msg, wp, lp);
+    WINE_TRACE("Calling guest proc 0x%lx(%p, %x, %lu, %lu).\n", (unsigned long)*guest_proc, hdlg,
+            msg, wp, lp);
     call->guest_proc = *guest_proc;
     call->dlg = (ULONG_PTR)hdlg;
     call->msg = msg;
